+++ GestorDeVentasApp/src/main/java/com/Hokkaido/GestorDeVentasApp/entidades/Sales_prod.java
@@ -8,27 +8,32 @@ public class Sales_prod {
 	@Id
 	private Long sale_id;
 	private Long product_id;
-	private Long quantity;
+	private int quantity;
 	
 	public Sales_prod() {
 		super();
 	}
+	
 	public Long getSale_id() {
 		return sale_id;
 	}
+
 	public void setSale_id(Long sale_id) {
 		this.sale_id = sale_id;
 	}
+
 	public Long getProduct_id() {
 		return product_id;
 	}
+
 	public void setProduct_id(Long product_id) {
 		this.product_id = product_id;
 	}
-	public Long getQuantity() {
+
+	public int getQuantity() {
 		return quantity;
 	}
-	public void setQuantity(Long quantity) {
+	public void setQuantity(int quantity) {
 		this.quantity = quantity;
 	}
